old_codebase/pyproject.toml
[project]
name = "civilization-game"
version = "0.1.0"
description = "An AI-driven civilization simulation game."
authors = [
    {name = "Your Name"}
]
readme = "README.md"
requires-python = ">=3.12"
package-mode = false
dependencies = [
    "flask>=3.0.0",
    "google-generativeai>=0.5.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0"
]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
